\name{get.ObjectClasses}
\alias{get.ObjectClasses}
\title{Returns a vector of classes for the objects in the user's workspace...}
\description{Returns a vector of classes for the objects in the user's workspace
; used by the GUI to retrieve a list of data frames in the workspace}
\keyword{internal}
