\name{get.VarTypes}
\alias{get.VarTypes}
\title{Returns a vector of variable types of the variables in the given df...}
\usage{get.VarTypes(df)}
\description{Returns a vector of variable types of the variables in the given df
; (the vector's labels are the variable names)}
\keyword{internal}
\arguments{\item{df}{data frame}}
