\name{updateMeltCodeBox}
\alias{updateMeltCodeBox}
\title{Updates the code preview box in the Melt tab...}
\usage{updateMeltCodeBox(code, id, measure, dataName)}
\description{Updates the code preview box in the Melt tab}
\keyword{internal}
\arguments{\item{code}{widget to update}
\item{id}{vector of ID variables}
\item{measure}{vector of measure variables}
\item{dataName}{character string of data frame}}
